refactor helpers, use new test helpers
